pyproject.toml
[tool.poetry]
name = "aphylogeo"
version = "0.1.0"
description = "A phylogenetic and geographic analysis tool"
authors = ["'Tahiri Lab'"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.0,<=3.11.5"
bio = "^1.5.9"
multiprocess = "^0.70.15"
pandas = "^2.1.1"
psutil = "^5.9.5"
pyyaml = "^6.0.1"
pytest = "^7.4.2"
numpy = "^1.26.0"
cython = "^3.0.2"
pymuscle5 = {git = "https://github.com/althonos/pymuscle5"}
robinson-foulds = "^1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "aphylogeo/"
